* Type : Update Description:     1. add Log message     2. update model class (OK)     3. update database (OK)     4. update controller (Going)

+++ src/edu/mcscheduling/model/Doctor.java
@@ -1,22 +1,21 @@
 package edu.mcscheduling.model;
 
-import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
 import android.content.ContentValues;
+import edu.mcscheduling.common.Logger;
 import edu.mcscheduling.common.StatusCode;
 import edu.mcscheduling.database.DatabaseDriver;
+import edu.mcscheduling.database.MsResultSet;
 import edu.mcscheduling.database.Transation;
 
 public class Doctor {
 	private DatabaseDriver db;
-	private boolean exist = false;
 	
 	public Doctor(DatabaseDriver db) {
 		if ( db != null ) {
 			this.db = db;
-			exist = true;
 		} 
 	}
 	
@@ -35,36 +34,36 @@ public class Doctor {
 				DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid),
 				DatabaseTable.Doctor.colDorNo);
 
-		ResultSet rs = db.select(sql);
+		MsResultSet rsVal = null;
+		rsVal = db.select(sql);
+		
 		int number = 1;
-		if ( rs == null ) {
+		if ( rsVal.rs == null ) {
 			return number;
 		} else {
 			try {
-				rs.next();
-				number = rs.getInt(1) + 1;
+				rsVal.rs.next();
+				number =  rsVal.rs.getInt(1) + 1;
 			} catch ( Exception e ) {
 				return number;
 			}
-
 		}
 		return number;
 	}
 	
 	public int addDoctor(final String userid,final String depName,final String dorName,final String telephone, 
 				final String jobTitle, final String history,final String subject,final String desc,final String picPath) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
 		
-			Object obj =
-				db.excuteTransation( new Transation() {
+		return db.excuteTransation( new Transation() {
+					@Override
+					public 	Integer execute(Object retValue) throws Exception {
+						int number = calcuteDoctorNumber(userid);
 				
-				@Override
-				public 	Integer execute() throws SQLException {
-		
-					int number = calcuteDoctorNumber(userid);
-					
-					String dorNo =  String.format("%05d", number);
-					
-					String sql = String.format("INSERT INTO %s (%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s) " +
+						String dorNo =  String.format("%05d", number);
+				
+						String sql = String.format("INSERT INTO %s (%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s) " +
 							"VALUES (%s,'%s','%s','%s','%s','%s','%s','%s','%s','%s','%s','%s')",
 							// Columns
 							DatabaseTable.Doctor.name,
@@ -95,21 +94,17 @@ public class Doctor {
 							new DateTime().getDateTime(),
 							//desc,
 							picPath);
-				
-					System.out.println(sql);
-					if ( db.insert(sql) < 0)  
-						throw new SQLException();
-					return number;
-				}
-			});
-			
-			return 0;
-			
+						System.out.println(sql);
+						return db.insert(sql);			
+					}
+			},null);	
 	}
 	
 	public int updateDoctor(String userid, String dorNo, String depName, String dorName, String telephone, 
 							String jobTitle, String history,String subject, String desc, String picPath) {
-
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
+		
 		String sql = String.format("UPDATE %s SET %s='%s',%s='%s',%s='%s'," +
 				"%s='%s',%s='%s',%s='%s',%s='%s',%s='%s'," +
 				"%s='%s',%s='%s'"+
@@ -131,43 +126,46 @@ public class Doctor {
 			DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid),
 			DatabaseTable.Doctor.colDorNo, dorNo);
 	
-		if ( db.update(sql) < 0 ) 
-			return StatusCode.WAR_REGISTER_FAIL();
-		return StatusCode.success;
+		return db.update(sql);
 	}
 	
 	public int deleteDoctor(String userid, String dorNo) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
+		else if ( dorNo == null || dorNo.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_DOCTOR_NUM_ERROR);
 		
 		String sql = String.format("DELETE FROM %s WHERE %s=%s AND %s='%s'", 
 						DatabaseTable.Doctor.name,
 						DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid),
 						DatabaseTable.Doctor.colDorNo, dorNo);
 		System.out.println(sql);
-		if ( db.delete(sql) < 0 ) 
-			return -1;
-		return StatusCode.success;
+		return db.delete(sql);
 	}
 	
 	public int deleteDoctorAll(String userid) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
+		
 		String sql = String.format("DELETE FROM %s WHERE %s=%s", 
 				DatabaseTable.Doctor.name,
 				DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid));
 		
-		if ( db.delete(sql) < 0 ) 
-			return -1;
-		return StatusCode.success;
+		return db.delete(sql);
 	}
 	
-	
 	private int getDoctorCount(String sql) {
-		ResultSet rs = db.select(sql);
+		MsResultSet rsVal = new MsResultSet();
+		rsVal = db.select(sql);
+		
 		int rowCount = 0;
-		if ( rs == null ) {
+		
+		if ( rsVal.status != StatusCode.success) {
 			return rowCount;
 		} else {
 			try {
-				rs.next();
-				rowCount = rs.getInt(1);
+				rsVal.rs.next();
+				rowCount = rsVal.rs.getInt(1);
 			} catch ( SQLException e ) {
 				rowCount = 0;
 			}
@@ -175,54 +173,59 @@ public class Doctor {
 		return rowCount;
 	}
 	
-	private ContentValues[] setResultSetToContentValues(int rowCount, String sql) {
-		ResultSet rs = db.select(sql);
+	private MsContentValues getDoctorContentValues(int rowCount, String sql) {
+		MsResultSet rsVal = null;
+		rsVal = db.select(sql);
 		
-		if ( rs == null ) {
-			return null;
-		}
-
-		ContentValues[] content = new ContentValues[rowCount];
+		if ( rsVal.status != StatusCode.success )
+			return new MsContentValues(rsVal.status);
+		
+		MsContentValues cvValue = new MsContentValues(StatusCode.success);
+		cvValue.cv = new ContentValues[rowCount];
+		
 		try {
 			int i = 0;
-			ResultSetMetaData meta = rs.getMetaData();
-			while (rs.next()) {
-				content[i] = new ContentValues();
+			ResultSetMetaData meta = rsVal.rs.getMetaData();
+			while (rsVal.rs.next()) {
+				cvValue.cv[i] = new ContentValues();
 				for ( int j=1; j<=meta.getColumnCount(); j++ ){
-					content[i].put(meta.getColumnName(j),rs.getString(meta.getColumnName(j)));
+					cvValue.cv[i].put(meta.getColumnName(j),rsVal.rs.getString(meta.getColumnName(j)));
 				}
 				i++;
 			}
-			rs.close();
-		}  catch (SQLException e) {
-			content = null;
+			rsVal.rs.close();
+		} catch (SQLException e) {
+			return new MsContentValues(Logger.e(this, StatusCode.ERR_GET_MEMBER_INFO_FAIL));
 		} catch (Exception e ) {
-			content = null;
+			return new MsContentValues(Logger.e(this, StatusCode.ERR_UNKOWN_ERROR));
 		}
-		return content;
+		return cvValue;
 	}
 	
-	
-	public ContentValues[]  getDoctor(String userid) {
+	public MsContentValues  getDoctor(String userid) {
+		if ( userid == null || userid.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_USERID_ERROR));
 		
 		String sql =String.format("SELECT COUNT(*) FROM %s WHERE %s=%s",
 				DatabaseTable.Doctor.name,
 				DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid));
 		
 		int rowCount = getDoctorCount(sql);
-		
-		if ( rowCount <= 0 )
-			return null;
+		if ( rowCount <= 0 ) 
+			return new MsContentValues(Logger.e(this, StatusCode.WAR_DOCTOR_NOT_SETTING));
 		
 		sql = String.format("SELECT * FROM %s WHERE %s=%s",
 					DatabaseTable.Doctor.name,
 					DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid));
 		
-		return setResultSetToContentValues(rowCount, sql);
+		return getDoctorContentValues(rowCount, sql);
 	}
 	
-
-	public ContentValues[]  getDoctorByDepName(String userid,String depName) {
+	public MsContentValues  getDoctorByDepName(String userid, String depName) {
+		if ( userid == null || userid.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_USERID_ERROR));
+		else if ( depName == null || depName.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_USERID_ERROR));
 		
 		String sql =String.format("SELECT COUNT(*) FROM %s WHERE %s=%s AND %s='%s'",
 				DatabaseTable.Doctor.name,
@@ -230,15 +233,14 @@ public class Doctor {
 				DatabaseTable.Doctor.colDepName, depName);
 		
 		int rowCount = getDoctorCount(sql);
-		
-		if ( rowCount <= 0 )
-			return null;
+		if ( rowCount <= 0 ) 
+			return new MsContentValues(Logger.e(this, StatusCode.WAR_DOCTOR_NOT_SETTING));
 		
 		sql = String.format("SELECT * FROM %s WHERE %s=%s AND %s='%s'",
 					DatabaseTable.Doctor.name,
 					DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid),
 					DatabaseTable.Doctor.colDepName, depName);
 		
-		return setResultSetToContentValues(rowCount, sql);
+		return getDoctorContentValues(rowCount, sql);
 	}
 }
