+++ src/edu/mcscheduling/model/DoctorSchedule.java
@@ -4,19 +4,19 @@ import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
+import edu.mcscheduling.common.Logger;
 import edu.mcscheduling.common.StatusCode;
 import edu.mcscheduling.database.DatabaseDriver;
+import edu.mcscheduling.database.MsResultSet;
 import edu.mcscheduling.database.Transation;
 import android.content.ContentValues;
 
 public class DoctorSchedule {
 	private DatabaseDriver db;
-	private boolean exist = false;
 	
 	public DoctorSchedule(DatabaseDriver db) {
 		if ( db != null ) {
 			this.db = db;
-			exist = true;
 		} 
 	}
 	
@@ -29,14 +29,13 @@ public class DoctorSchedule {
 
 	public int addDoctorSchedule(final String userid, final String dorNo,final String depName, 
 			final String schYear, final String schMonth, final String schedule,final String desc) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
 		
-		
-		Object obj =
-				db.excuteTransation( new Transation() {
-				
-				@Override
-				public 	Integer execute() throws SQLException {
-					String sql = String.format("INSERT INTO %s (%s,%s,%s,%s,%s,%s,%s,%s) " +
+		return db.excuteTransation( new Transation() {		
+					@Override
+					public 	Integer execute(Object retValue) throws Exception {
+						String sql = String.format("INSERT INTO %s (%s,%s,%s,%s,%s,%s,%s,%s) " +
 							"VALUES (%s,'%s','%s','%s','%s','%s','%s','%s')",
 							// Columns
 							DatabaseTable.DoctorSchedule.name,
@@ -59,75 +58,74 @@ public class DoctorSchedule {
 							userid,
 							new DateTime().getDateTime());
 							//desc);
-
-					if ( db.insert(sql) < 0 )
-						throw new SQLException();
-					return StatusCode.success;
+					return db.insert(sql);
 				}
-			});
-			
-			return 0;
+			}, null);
 	}
 	
 	public int updateDoctorSchedule(String userid, String dorNo, String depName,
 		String schYear, String schMonth, String schedule, String desc) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
 		
 		String sql = String.format("UPDATE %s SET %s=%s,%s='%s',%s='%s',%s='%02d'," +
-										"%s='%s',%s='%s',%s='%s'" +
-										"WHERE %s=%s AND %s='%s' AND %s='%s' AND %s='%s'" ,
-			// Table
-			DatabaseTable.DoctorSchedule.name, 
-			// Value
-			DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
-			DatabaseTable.DoctorSchedule.colDepName, depName,
-
-			DatabaseTable.DoctorSchedule.colSchYear, schYear,
-			DatabaseTable.DoctorSchedule.colSchMonth, Integer.valueOf(schMonth),
-			DatabaseTable.DoctorSchedule.colSchedule, schedule,
-			DatabaseTable.DoctorSchedule.colUpdateID, userid,
-			DatabaseTable.DoctorSchedule.colUpdateDate, new DateTime().getDateTime(),
-			//DatabaseTable.DoctorSchedule.colDesc, "NULL",
-			// WHERE
-			DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
-			DatabaseTable.DoctorSchedule.colDorNo, dorNo,
-			DatabaseTable.DoctorSchedule.colSchYear, schYear,
-			DatabaseTable.DoctorSchedule.colSchMonth, schMonth);
-
-		if ( db.update(sql) < 0 ) 
-			return StatusCode.WAR_REGISTER_FAIL();
-		return StatusCode.success;
+									"%s='%s',%s='%s',%s='%s'" +
+									"WHERE %s=%s AND %s='%s' AND %s='%s' AND %s='%s'" ,
+									// Table
+									DatabaseTable.DoctorSchedule.name, 
+									// Value
+									DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
+									DatabaseTable.DoctorSchedule.colDepName, depName,
+									DatabaseTable.DoctorSchedule.colSchYear, schYear,
+									DatabaseTable.DoctorSchedule.colSchMonth, Integer.valueOf(schMonth),
+									DatabaseTable.DoctorSchedule.colSchedule, schedule,
+									DatabaseTable.DoctorSchedule.colUpdateID, userid,
+									DatabaseTable.DoctorSchedule.colUpdateDate, new DateTime().getDateTime(),
+									//DatabaseTable.DoctorSchedule.colDesc, "NULL",
+									// WHERE
+									DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
+									DatabaseTable.DoctorSchedule.colDorNo, dorNo,
+									DatabaseTable.DoctorSchedule.colSchYear, schYear,
+									DatabaseTable.DoctorSchedule.colSchMonth, schMonth);
+		return db.update(sql);
 	}
 		
 	public int deleteDoctorSchedule(String userid, String dorNo) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
+		else if ( dorNo == null || dorNo.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_DOCTOR_NUM_ERROR);
+		
 		String sql = String.format("DELETE FROM %s WHERE %s=%s AND %s='%s'", 
 						DatabaseTable.DoctorSchedule.name,
 						DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
 						DatabaseTable.DoctorSchedule.colDorNo, dorNo);
 		
-		if ( db.delete(sql) < 0 ) 
-			return -1;
-		return StatusCode.success;
+		return db.delete(sql);
 	}
 	
 	public int deleteDoctorScheduleAll(String userid) {
+		if ( userid == null || userid.isEmpty() )
+			return Logger.e(this, StatusCode.PARM_USERID_ERROR);
+		
 		String sql = String.format("DELETE FROM %s WHERE %s=%s", 
 						DatabaseTable.DoctorSchedule.name,
 						DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid));
-		
-		if ( db.delete(sql) < 0 ) 
-			return -1;
-		return StatusCode.success;
+		return db.delete(sql);
 	}
 	
 	private int getDoctorScheduleCount(String sql) {
-		ResultSet rs = db.select(sql);
+		MsResultSet rsVal = new MsResultSet();
+		rsVal = db.select(sql);
+		
 		int rowCount = 0;
-		if ( rs == null ) {
+		
+		if ( rsVal.status != StatusCode.success) {
 			return rowCount;
 		} else {
 			try {
-				rs.next();
-				rowCount = rs.getInt(1);
+				rsVal.rs.next();
+				rowCount = rsVal.rs.getInt(1);
 			} catch ( SQLException e ) {
 				rowCount = 0;
 			}
@@ -135,54 +133,62 @@ public class DoctorSchedule {
 		return rowCount;
 	}
 	
-	private ContentValues[] setResultSetToContentValues(int rowCount, String sql) {
-		ResultSet rs = db.select(sql);
+	private MsContentValues getDoctorScheduleContentValues(int rowCount, String sql) {
+		MsResultSet rsVal = null;
+		rsVal = db.select(sql);
 		
-		if ( rs == null ) {
-			return null;
-		}
-
-		ContentValues[] content = new ContentValues[rowCount];
+		if ( rsVal.status != StatusCode.success )
+			return new MsContentValues(rsVal.status);
+		
+		MsContentValues cvValue = new MsContentValues(StatusCode.success);
+		cvValue.cv = new ContentValues[rowCount];
+		
 		try {
 			int i = 0;
-			ResultSetMetaData meta = rs.getMetaData();
-			while (rs.next()) {
-				content[i] = new ContentValues();
+			ResultSetMetaData meta = rsVal.rs.getMetaData();
+			while (rsVal.rs.next()) {
+				cvValue.cv[i] = new ContentValues();
 				for ( int j=1; j<=meta.getColumnCount(); j++ ){
-					content[i].put(meta.getColumnName(j),rs.getString(meta.getColumnName(j)));
+					cvValue.cv[i].put(meta.getColumnName(j),rsVal.rs.getString(meta.getColumnName(j)));
 				}
 				i++;
 			}
-			rs.close();
-		}  catch (SQLException e) {
-			content = null;
+			rsVal.rs.close();
+		} catch (SQLException e) {
+			return new MsContentValues(Logger.e(this, StatusCode.ERR_GET_MEMBER_INFO_FAIL));
 		} catch (Exception e ) {
-			content = null;
+			return new MsContentValues(Logger.e(this, StatusCode.ERR_UNKOWN_ERROR));
 		}
-		return content;
+		return cvValue;
 	}
 	
-	
-	
-	
-	public ContentValues[]  getDoctorSchedule(String userid) {
+	public MsContentValues  getDoctorSchedule(String userid) {
+		if ( userid == null || userid.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_USERID_ERROR));
+		
 		String sql = String.format("SELECT COUNT(*) FROM %s WHERE %s=%s", 
 				DatabaseTable.DoctorSchedule.name,
 				DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid));
 		
 		int rowCount = getDoctorScheduleCount(sql);
-		
-		if ( rowCount <= 0 )
-			return null;
+		if ( rowCount <= 0 ) 
+			return new MsContentValues(Logger.e(this, StatusCode.WAR_DOCTORSCHEDULE_NOT_SETTING));
 		
 		sql = String.format("SELECT * FROM %s WHERE %s=%s",
 				DatabaseTable.Doctor.name,
 				DatabaseTable.Doctor.colHospitalNo, getHospitalNoSQL(userid));
 	
-		return setResultSetToContentValues(rowCount, sql);
+		return getDoctorScheduleContentValues(rowCount, sql);
 	}
 
-	public ContentValues[] getDoctorScheduleByDepName_AND_DorNo_AND_SchYear_SchMonth(String userid, String depName, String dorNo,int year, int month) {
+	public MsContentValues getDoctorScheduleByDepName_AND_DorNo_AND_SchYear_SchMonth(String userid, String depName, String dorNo,int year, int month) {
+		if ( userid == null || userid.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_USERID_ERROR));
+		else if ( depName == null || depName.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_DEPART_NAME_ERROR));
+		else if ( dorNo == null || dorNo.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_DOCTOR_NUM_ERROR));
+		
 		String sql = String.format("SELECT COUNT(*) FROM %s WHERE %s=%s AND %s='%s' AND %s='%s' AND %s='%d' AND %s='%02d'",
 				DatabaseTable.DoctorSchedule.name,
 				DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
@@ -192,9 +198,8 @@ public class DoctorSchedule {
 				DatabaseTable.DoctorSchedule.colSchMonth,Integer.valueOf(month));
 		
 		int rowCount = getDoctorScheduleCount(sql);
-		
-		if ( rowCount <= 0 )
-			return null;
+		if ( rowCount <= 0 ) 
+			return new MsContentValues(Logger.e(this, StatusCode.WAR_DOCTORSCHEDULE_NOT_SETTING));
 		
 		sql = String.format("SELECT * FROM %s WHERE %s=%s AND %s='%s' AND %s='%s' AND %s='%d' AND %s='%02d'",
 				DatabaseTable.DoctorSchedule.name,
@@ -204,10 +209,19 @@ public class DoctorSchedule {
 				DatabaseTable.DoctorSchedule.colSchYear,year,
 				DatabaseTable.DoctorSchedule.colSchMonth,Integer.valueOf(month));
 	
-		return setResultSetToContentValues(rowCount, sql);
+		return getDoctorScheduleContentValues(rowCount, sql);
 	}
 
-	public ContentValues[] getDoctorScheduleByDorNo_ShcYear_ShcMonth(String userid, String dorNo, String year, String month) {
+	public MsContentValues getDoctorScheduleByDorNo_ShcYear_ShcMonth(String userid, String dorNo, String year, String month) {
+		if ( userid == null || userid.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_USERID_ERROR));
+		else if ( dorNo == null || dorNo.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_DOCTOR_NUM_ERROR));
+		else if ( year == null || year.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_SCH_YEAR_ERROR));
+		else if ( month == null || month.isEmpty() )
+			return new MsContentValues(Logger.e(this, StatusCode.PARM_SCH_MONTH_ERROR));
+		
 		String sql = String.format("SELECT COUNT(*) FROM %s WHERE %s=%s AND %s='%s' AND %s='%s' AND %s='%02d'",
 				DatabaseTable.DoctorSchedule.name,
 				DatabaseTable.DoctorSchedule.colHospitalNo, getHospitalNoSQL(userid),
@@ -217,8 +231,8 @@ public class DoctorSchedule {
 		
 		int rowCount = getDoctorScheduleCount(sql);
 		
-		if ( rowCount <= 0 )
-			return null;
+		if ( rowCount <= 0 ) 
+			return new MsContentValues(Logger.e(this, StatusCode.WAR_DOCTORSCHEDULE_NOT_SETTING));
 		
 		sql = String.format("SELECT * FROM %s WHERE %s=%s AND %s='%s' AND %s='%s' AND %s='%02d'",
 				DatabaseTable.DoctorSchedule.name,
@@ -227,8 +241,6 @@ public class DoctorSchedule {
 				DatabaseTable.DoctorSchedule.colSchYear, year,
 				DatabaseTable.DoctorSchedule.colSchMonth, Integer.valueOf(month));
 	
-		return setResultSetToContentValues(rowCount, sql);
+		return getDoctorScheduleContentValues(rowCount, sql);
 	}
-	
-	
 }
